Updating domain of joint PMF for discrete RVs
The original domain could work when Val(X) and Val(Y) are both subsets of the real numbers, and if they aren't then in most cases they can be mapped to real numbers; either way, it seems odd to let p_{XY} be defined for (x,y) not in Val(X) x Val(Y). One consequence of doing so is that such (x,y) could be assigned nonzero probability, since we only require that the sum over Val(X) x Val(Y) equals 1!).