Fixing log of phi issue
Updating max expressions to be more rigorous (since the log of phi(c) isn't phi(c)). Used theta_c to match the lecture slides and thetas used later in this lecture notes entry (which were originally used without introduction).

Alternatively we could define phi(c) to be a log-likelihood, but we should make that more clear if we do.