Clarify I-Maps
Used to say, "Let I(p) denote the set of all CONDITIONAL independencies that hold for a distribution p" but the example given is just a regular ol' independence rather than an explicit conditional independence. So I changed it to say "the set of all independencies" which aligns with the example and with the PGM textbook.