improved grammar/definition of d-separation
"X affects Y in addition to Z" can mean either "X affects Y and Z" or "X and Z affect Y"; I changed the wording to eliminate this ambiguity

The definition of d-ambiguity didn't say what Q and W were; I added that they contained vertices from G. "they" was used to refer to Q and W, but the reader could've easily thought that it referred instead to O, the variables that are observed. So I got rid of the ambiguity there, too.