Weaken claim about CDF of X (#182)
There are some events (sets of outcomes in Omega) that can't be distinguished just by using the CDF on the random variable X. For example, if your X(omega) is "sum of number of heads over 2 (non-independent) coin flips", then you can't use the CDF of X to compute the probability of the event containing the single outcome "first get a head, then get a tails". In this example, you can only use the CDF to compute the probability of the event containing the two outcomes "first head then tails" and "first tails then heads".